+++ src/main/php/ioc/FileBasedLoggerProvider.php
@@ -1,4 +1,5 @@
 <?php
+declare(strict_types=1);
 /**
  * This file is part of stubbles.
  *
@@ -9,6 +10,7 @@
  */
 namespace stubbles\log\ioc;
 use stubbles\ioc\InjectionProvider;
+use stubbles\log\Logger;
 use stubbles\log\appender\FileLogAppender;
 /**
  * Injection provider for logger instances with a file appender.
@@ -45,7 +47,7 @@ class FileBasedLoggerProvider implements InjectionProvider
      * @Named{logPath}('stubbles.log.path')
      * @Named{fileMode}('stubbles.log.filemode')
      */
-    public function __construct(LoggerProvider $loggerProvider, $logPath, $fileMode = 0700)
+    public function __construct(LoggerProvider $loggerProvider, string $logPath, int $fileMode = 0700)
     {
         $this->loggerProvider = $loggerProvider;
         $this->logPath        = $logPath;
@@ -58,7 +60,7 @@ class FileBasedLoggerProvider implements InjectionProvider
      * @param   string  $name  optional
      * @return  \stubbles\log\Logger
      */
-    public function get(string $name = null)
+    public function get(string $name = null): Logger
     {
         $logger = $this->loggerProvider->get($name);
         if (!$logger->hasLogAppenders()) {